Fix bordering

+++ picombiner.rb
@@ -8,7 +8,7 @@
 #
 # Usage:
 # 
-# `$ ruby picombiner.rb [dir]`
+# `$ ruby picombiner.rb [dir] [tile1] [tile2]`
 #
 
 require 'fileutils'
@@ -16,32 +16,30 @@ require 'fileutils'
 FILE_RULE = "*.{JPG,jpg,PNG,png,TIF,tif}"
 COLORS = ['white', 'black']
 BORDER_FACTOR = 0.0125
-DIRECTIONS = {:horizontal => "+append", :vertical => "-append"}
 
 @@in_dir  = File.directory?(ARGV[0]) ? ARGV[0] : File.expand_path(File.join(File.dirname(__FILE__), ARGV[0]))
 @@tmp_dir = File.join(@@in_dir, 'tmp')
 @@out_dir = File.join(@@in_dir, 'out')
 
+@@tiles = ARGV[1..ARGV.size-1]
+raise "tile not given" if @@tiles.empty?
+
 @@files = Dir[File.join(@@in_dir, FILE_RULE)]
 raise "No files found!" if @@files.empty?
 
-@@base_border = (`identify -format "%w" #{@@files.first}`.to_f * BORDER_FACTOR).to_i
+@@base_border = (`identify -format "%w" #{@@files.first}`.to_f * BORDER_FACTOR).to_i / 2
 
 FileUtils.rm_rf(@@out_dir)
 FileUtils.mkdir_p(@@out_dir)
 
 def files_in(dir)
-  Dir[File.join(dir, FILE_RULE)]
+  Dir[File.join(dir, FILE_RULE)].sort
 end
 
-def output_file(direction, color)
+def output_file(tile, color)
   first_file = Dir[File.join(@@in_dir, FILE_RULE)].sort.first
-  file_name = [File.basename(first_file, ".*"), direction[0], color[0], File.extname(first_file)].join
+  file_name = [File.basename(first_file, ".*"), '-', tile, color[0], File.extname(first_file)].join
   return File.join(@@out_dir, file_name)
-end
-
-def picture_border(idx, total)
-  return (idx == 0 || idx == total-1) ? @@base_border : 0
 end
 
 def exec(cmd)
@@ -49,36 +47,20 @@ def exec(cmd)
   `#{cmd}`
 end
 
-def border_im_operator(border, direction)
-  case direction 
-    when :horizontal then "#{border}x#{@@base_border}"
-    when :vertical then "#{@@base_border}x#{border}"
-    else raise "Invalid IM direction"
-  end
+def montage(tile, color)
+  file_list = files_in(@@in_dir).join(' ')
+  exec "montage -background #{color} -geometry +#{@@base_border}+#{@@base_border} -tile #{tile} #{file_list} #{output_file(tile, color)}"
 end
 
-def border_each!(direction, color)
-  @@files.each_with_index do |f, idx|
-    file_path = File.join(@@tmp_dir, File.basename(f))
-    current_border = picture_border(idx, @@files.size)
-    exec "convert -bordercolor #{color} -border #{border_im_operator(current_border, direction)} '#{f}' '#{file_path}'"
-  end
+def fix_border(tile, color)
+  file = output_file(tile, color)
+  exec "convert -bordercolor #{color} -border #{@@base_border}x#{@@base_border} #{file} #{file}"
 end
 
-def combine!(direction, color)
-  file_list = Dir[File.join(@@tmp_dir, FILE_RULE)].join(' ')
-  exec "convert #{file_list} #{DIRECTIONS[direction]} '#{output_file(direction, color)}'"
-end
-
-DIRECTIONS.keys.each do |direction|
-  COLORS.each do |color|
-    puts "*** #{direction} bordering with #{color} #{@@base_border}px ***" 
-    begin
-      FileUtils.mkdir_p(@@tmp_dir)
-      border_each!(direction, color)
-      combine!(direction, color)
-    ensure
-      FileUtils.rm_rf(@@tmp_dir)
-   end
+COLORS.each do |color|
+  @@tiles.each do |tile|
+    puts "*** #{tile} with #{color} #{@@base_border*2}px ***" 
+    montage(tile, color)
+    fix_border(tile, color)
   end
 end
